Move bookmarks item component
https://youtrack.weseek.co.jp/issue/GW-7861
- Remove rendering bookmark items from Bookmarks component
- Move BookmarkItem component implementation to BookmarkFolderTree
- Fix tooltip text of bookmark item
- Adjust styling of bookmark item
- Adjust Bookmark folder display for guest user
- Adjust render bookmark item when no bookmarked pages
- Remove unused definition of virtual bookmarks
- Remove folder from bookmark Schema
- Update return value of bookmark folder list route
